% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_schema_form.R
\name{jsonSchemaFormInput}
\alias{jsonSchemaFormInput}
\title{Shiny Input Form from JSON Schema}
\usage{
jsonSchemaFormInput(
  inputId,
  schema,
  default = NA,
  setValueOnChange = FALSE,
  ...
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{schema}{An R object using JSON schema keywords that configures
the form fields.}

\item{default}{Initial value.}

\item{setValueOnChange}{Whether \code{input} slot be updated as data is entered
into form making submit button decorative.}

\item{...}{Additional arguments passed to react-jsonschema-form \verb{<Form />}.}
}
\description{
Allows you to use a json schema to create a form-style shiny input.
}
\details{
Wraps \href{https://rjsf-team.github.io/react-jsonschema-form/docs/}{react-jsonschema-form}.
}
