% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcr-package.R
\docType{package}
\name{srcr-package}
\alias{srcr-package}
\title{srcr: Simplify Connections to Database Sources}
\description{
Connecting to databases requires boilerplate code to specify connection parameters and to set up sessions properly with the DBMS. This package provides a simple tool to fill two purposes: abstracting connection details, including secret credentials, out of your source code and managing configuration for frequently-used database connections in a persistent and flexible way, while minimizing requirements on the runtime environment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/baileych/srcr}
  \item Report bugs at \url{https://github.com/baileych/srcr/issues}
}

}
\author{
\strong{Maintainer}: Charles Bailey \email{baileyc@chop.edu}

Authors:
\itemize{
  \item Hanieh Razzaghi \email{razzaghih@chop.edu}
}

}
\keyword{internal}
