% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_schema_form.R
\docType{data}
\name{VALID_FORM_PROPS}
\alias{VALID_FORM_PROPS}
\title{onSubmit',}
\format{
An object of class \code{character} of length 38.
}
\usage{
VALID_FORM_PROPS
}
\description{
onSubmit',
}
\keyword{datasets}
